+++ Changelog.md
@@ -13,6 +13,8 @@ Breaking Changes:
 * Raise on usage of metadata on suite-level scopes. (Phil Pirozhkov, #2849)
 * Raise an error when `fail_fast` is configured with
   an unsupported value. (Phil Pirozhkov, #2849)
+* Remove deprecated access to an example group's metadata through the example.
+  (Phil Pirozhkov, #2851)
 
 Enhancements: